plugin: add picom API headers
And add a pkgconfig file so they can be found.
