+++ include/meson.build
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: MPL-2.0
+# Copyright (c) Yuxuan Shui <yshuiv7@gmail.com>
+subdirs('picom')
