+++ include/picom/meson.build
@@ -1,0 +1,8 @@
+# SPDX-License-Identifier: MPL-2.0
+# Copyright (c) Yuxuan Shui <yshuiv7@gmail.com>
+
+api_headers = [
+  'api.h'
+  'backend.h'
+]
+install_headers(api_headers, subdir: 'picom')
