+++ src/backend/backend.c
@@ -4,13 +4,14 @@
 #include <xcb/sync.h>
 #include <xcb/xcb.h>
 
+#include <picom/types.h>
+
 #include "backend/backend.h"
 #include "common.h"
 #include "compiler.h"
 #include "config.h"
 #include "log.h"
 #include "region.h"
-#include "types.h"
 #include "win.h"
 #include "x.h"
 
@@ -21,10 +22,10 @@ static struct backend_info {
 	bool can_present;
 } *backend_registry = NULL;
 
-bool PICOM_PUBLIC_API backend_register(uint64_t major, uint64_t minor, const char *name,
-                                       struct backend_base *(*init)(session_t *ps,
-                                                                    xcb_window_t target),
-                                       bool can_present) {
+PICOM_PUBLIC_API bool
+backend_register(uint64_t major, uint64_t minor, const char *name,
+                 struct backend_base *(*init)(session_t *ps, xcb_window_t target),
+                 bool can_present) {
 	if (major != PICOM_BACKEND_MAJOR) {
 		log_error("Backend %s has incompatible major version %" PRIu64
 		          ", expected %lu",
