+++ src/renderer/damage.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "types.h"
+#include <picom/types.h>
 
 typedef struct pixman_region32 region_t;
 struct layout;
